Move DistinguishedName into verify module
